feat: 플레이어 데미지 주는 기능 추가 중 +이슈: 파티클 공격에서 데미지가 연속으로 들어가는 이슈 발생 +이슈: 보스 물기 공격 데미지 전달 않됨

+++ Royal Rescue/Assets/Scripts/Boss/Skillinfo/ScriptableObjectScripts/BasicAttackScriptableObject.cs
@@ -7,6 +7,6 @@ public class BasicAttackScriptableObject : ScriptableObject
 {
     public Transform transform;
     public float BasicAttackDistance = 0;
-    public float damage = 0;
+    public int damage = 0;
     public float subSequenceDelay;
 }
