+++ Royal Rescue/Assets/Scripts/Boss/Skillinfo/ScriptableObjectScripts/FlameAttackScriptableObject.cs
@@ -6,7 +6,7 @@ public class FlameAttackScriptableObject : ScriptableObject
     public GameObject flameObj;
     public float flameAttackDistance = 0;
     public int flameCount;
-    public float damage = 0;
+    public int damage = 0;
     public float flameSpeed = 0;
     public float subSequenceDelay;
 }
